Node "Item" panel update.
This adds:
* "Replace node..." menu to the panel
* "Re-Create node" to the panel
